LayoutTests/accessibility/aria-modal-in-aria-hidden.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/accessibility-helper.js"></script>
<script src="../resources/js-test.js"></script>
</head>
<body>

<div id="bg">
<p id="bgContent">Other page content with <a href="#">a dummy focusable element</a></p>
</div>

<div id="container">
    <div role="dialog" aria-modal="false" id="box">
        <h3 id="myDialog">Just an example.</h3>
    </div>
</div>

<script>
var output = "This tests that when something is aria-modal inside an aria-hidden it is ignored.\n\n";

if (window.accessibilityController) {
    window.jsTestIsAsync = true;
    output += expect("accessibilityController.accessibleElementById('bgContent').isIgnored", "false")

    document.getElementById("box").setAttribute("aria-modal", "true");
    setTimeout(async function() {
        output += await expectAsync("!accessibilityController.accessibleElementById('bgContent')", "true");

        document.getElementById("container").setAttribute("aria-hidden", "true");
        output += await expectAsync("accessibilityController.accessibleElementById('bgContent')?.isIgnored", "false")

        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
